refactor icons dir
