+++ src/App/Playlist/LoadedPlaylist/Sidebar/Icons/DevicesIcon/Devices.scss
