+++ src/App/Playlist/LoadedPlaylist/Sidebar/Icons/DevicesIcon/index.js
