+++ src/App/Playlist/LoadedPlaylist/Sidebar/Icons/LockIcon/Lock.scss
