+++ src/App/Playlist/LoadedPlaylist/Sidebar/Icons/LockIcon/index.js
