+++ src/App/Playlist/LoadedPlaylist/Sidebar/Icons/SearchIcon/Search.scss
