+++ src/App/Playlist/LoadedPlaylist/Sidebar/Icons/SearchIcon/index.js
