+++ src/App/Playlist/LoadedPlaylist/Sidebar/Icons/ShuffleIcon/Shuffle.scss
