+++ src/App/Playlist/LoadedPlaylist/Sidebar/Icons/ShuffleIcon/index.js
